added error handling

+++ ABZFlowers/backend/routes/products.js
@@ -33,7 +33,7 @@ const storage = multer.diskStorage({
 
 //multer will try to extract a single image
 router.post("", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
-  console.log("Jooj");
+
   const url = req.protocol + '://' + req.get("host");
   //you're calling this from the models folder product.js, you gave its name as'Product'
   const product = new Product({
@@ -43,6 +43,7 @@ router.post("", checkAuth, multer({ storage: storage }).single("image"), (req, r
     quantity: req.body.quantity,
     price: req.body.price,
     currency: req.body.currency,
+    creator: req.userData.userId
   }); //this object is being managed by mongoose, you can save the objects created here directly on monngoDB
   product.save().then(createdProduct => {
     res.status(201).json({
@@ -52,6 +53,11 @@ router.post("", checkAuth, multer({ storage: storage }).single("image"), (req, r
         id: createdProduct._id //override to _id property
       }
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Creating a post failed!"
+    })
   }); //mongoose function, will create a collection called the same name as the model but plural
 
   // We dont use next() jere because res is already sending a response
@@ -71,18 +77,26 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
     currency: req.body.currency,
     price: req.body.price,
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    creator: req.userData.userId
   })
   console.log(product);
   // Product.updateOne({ _id: req.params.id }, product).then(result => {
   //   res.status(200).json({ message: "Update successful!" });
   // })
-  Product.findOneAndUpdate({ _id: req.params.id }, { $set: product }, null).then(result => {
-    // if(err){
-    //   return res.status(400).json({error:'You are not authorized'})
-    // }
-    res.status(200).json({ message: "Update successful!" });
-    console.log('deubom');
+  Product.updateOne({ _id: req.params.id, creator: req.userData.userId }, product).then(result => {
+    console.log(result);
+    if (result.modifiedCount > 0){
+      res.status(200).json({ message: "Update successful!" });
+    }
+    else {
+     res.status(401).json({ message: "You are not authorized!" });
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Couldn't update post"
+    })
   });
 });
 
@@ -107,6 +121,11 @@ router.get("", (req, res, next) => {
         products: fetchedProducts,
         maxProducts: count
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching posts failed!"
+      })
     });
   //200 is sucess
 
@@ -119,14 +138,29 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({ message: 'Product not found!' })
     }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching post failed!"
+    })
   });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Product.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result)
+  Product.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.modifiedCount > 0){
+      res.status(200).json({ message: "Deletion successful!" });
+    }
+    else {
+      res.status(401).json({ message: "You are not authorized!" });
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Deleting post failed!"
+    })
   });
-  res.status(200).json({ message: "Product deleted!" });
+
 });
 
 module.exports = router;
